requirements 2.txt
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
PyPDF2>=3.0.0
openpyxl>=3.1.2
python-dateutil>=2.8.2 